refactor : 컨트롤러에 있는 메서드 FileInput 으로 이동
